+++ lib/grpc_web.dart
@@ -14,29 +14,12 @@
 // limitations under the License.
 
 export 'src/auth/auth.dart'
-    show
-        BaseAuthenticator,
-        HttpBasedAuthenticator,
-        ComputeEngineAuthenticator,
-        JwtServiceAccountAuthenticator;
+    show HttpBasedAuthenticator, JwtServiceAccountAuthenticator;
 
-export 'src/client/call.dart' show ClientCall;
+export 'src/client/call.dart' show MetadataProvider, CallOptions;
 
-export 'src/client/client.dart' show Client;
 export 'src/client/common.dart' show Response, ResponseStream, ResponseFuture;
-export 'src/client/method.dart' show ClientMethod;
-export 'src/client/options.dart'
-    show
-        defaultIdleTimeout,
-        BackoffStrategy,
-        defaultBackoffStrategy,
-        ChannelOptions,
-        MetadataProvider,
-        CallOptions;
 
-export 'src/client/transport/transport.dart';
 export 'src/client/web_channel.dart' show GrpcWebClientChannel;
 
 export 'src/shared/status.dart' show StatusCode, GrpcError;
-export 'src/shared/streams.dart' show GrpcHttpEncoder, GrpcHttpDecoder;
-export 'src/shared/timeout.dart' show toTimeoutString, fromTimeoutString;
